즐겨찾기 FavoriteDao 신설 + insertFavorite·deleteFavorite 이관(HugesoInfoDao→FavoriteDao)
분산돼 있던 즐겨찾기 로직을 표준 <도메인>/model/Dao 구조로 응집하는 첫 단계.

- favorite/model/FavoriteDao.java 신설(형제 DAO 패턴: db private·getConnection+finally·탭/K&R)
- HugesoInfoDao의 insertFavorite(FavoriteDto)·deleteFavorite(m_num,h_num) 2개를
  FavoriteDao로 이동(SQL·시그니처·로직 동일). 이후 죽은 import favorite.model.FavoriteDto 제거
- 호출처 갱신: insertfavorite.jsp·deletefavorite.jsp의 import·인스턴스를 FavoriteDao로
  (isLogin·isPost·checkCsrf·세션 m_num 강제 등 보안 가드 불변)

javac EXIT=0.
